Split tags into new module.
