+++ cirq/ops/raw_types.py
@@ -609,24 +609,6 @@ class TaggedOperation(Operation):
                                                   atol=atol)
 
 
-class VirtualTag():
-    """A TaggedOperation tag indicating that the operation is virtual.
-
-    Operations marked with this tag are presumed to have zero duration of their
-    own, although they may have a non-zero duration if run in the same Moment
-    as a non-virtual operation.
-    """
-
-    def __eq__(self, other):
-        return isinstance(other, VirtualTag)
-
-    def __str__(self):
-        return '<virtual>'
-
-    def __repr__(self):
-        return str(self)
-
-
 @value.value_equality
 class _InverseCompositeGate(Gate):
     """The inverse of a composite gate."""
